removed useless code, made look nice, added code
Some old code was removed, This old code was used for another project
involving a heath+humidity sendor and bluetooth. The code should run
smoothly now. I also orderd the code better so its prettyer to look at
and easier to find errors in the future. there is also some code added
in the bluetooth section, but this will (most probably) change in the
future.